sensor reading library updates

+++ Boards/sensor_boards/read_sensors.h
@@ -1,7 +1,7 @@
 #ifndef HEADER_FILE_SENSORS
 #define HEADER_FILE_SENSORS
 
-int read_sensors(int *pres, int *temp);
+int read_sensors(uint16_t *pres, int16_t *temp);
 int READ_OK = 0;
 int READ_FAIL = 1;
